Folder swap try 2.
Summary:
This seems a giant diff, but it actually does something simple, it performs the final folder swap for ProjectariaTools:
1. `projectaria_tools` -> `projectaria_tools_gen1_legacy`
2. `projectaria_tools_gen2` -> `projectaria_tools`.

For most BUCK target, it automatically depend on gen2 so there is nothing that we need to do. For a few targets that explicitly depends on gen1 or gen2, this diff fixes their BUCK deps target.

Reviewed By: kongchen1992

Differential Revision: D82388571

fbshipit-source-id: 840f128b851483f6ae745424112f53e475274ca0